# Merge ivar guards on JIT when an ISeq has multiple ivar accesses