```
$ benchmark-driver -v --rbenv 'before --jit;after --jit' benchmark.yml --repeat-count=24 --output=all
before --jit: ruby 2.8.0dev (2020-07-03T16:52:54Z master ccb2e6eaa5) +JIT [x86_64-linux]
after --jit: ruby 2.8.0dev (2020-07-03T23:58:06Z ivar-guard f3d3691c74) +JIT [x86_64-linux]
Calculating -------------------------------------
                                 before --jit           after --jit
Optcarrot Lan_Master.nes     71.3624688269922      70.9239548083316 fps
                             72.6359580311016      77.1039753707283
                             74.1993512788708      80.0204069821405
                             75.3320353557776      81.7604440781847
                             75.8100844520402      82.6924317730600
                             78.3256414692381      83.1069968129562
                             79.0091501069809      84.1156234337417
                             79.1959373760543      84.5740177219342
                             79.3106606484137      85.9564085474114
                             79.5170038163073      86.4532950598815
                             79.8300039575979      86.6334184285854
                             80.1353874907729      87.8352304838773
                             80.1378791759272      88.3262571688172
                             80.7475478759448      88.8220596131538
                             81.1574370642460      89.1693182394324
                             81.4321261692806      89.4351081801636
                             81.6494191482392      89.8590686370407
                             81.9902704472289      89.8956141186323
                             82.5367512828839      90.3904587601945
                             82.5612412463655      90.6481734159511
                             82.8173726568623      91.0229514349186
                             83.0947368065651      91.4468334000509
                             83.1425803843619      91.5750654820674
                             83.2655243563682      93.0285740785746
```

The technique has been implemented in other implementations like RTL-MJIT and TruffleRuby. Credit goes to the precedents.